Add safer update for URLs
Split the save method into Add and Replace methods. If the user tries to Add a resource that already exists, they get a conflict error. They can still update the resource via PUT method, but this would need to be deliberate, so it avoids someone accidentally overwriting a URL that is in use.